refactor string utilities to separate file
